moving commands to separate files
