+++ src/commands.rs
@@ -1,0 +1,2 @@
+pub(crate) mod append_event;
+pub(crate) mod search;
